+++ src/main/java/calculator/Plus.java
@@ -46,7 +46,7 @@ public final class Plus extends Operation
    * @param r The second integer that should be added to the first
    * @return The integer that is the result of the addition
    */
-  public int op(int l, int r) {
+  public double op(double l, double r) {
   	return (l+r);
   }
 }
