+++ src/main/java/calculator/Times.java
@@ -45,6 +45,6 @@ public final class Times extends Operation
    * @param r The second integer that should be multiplied with the first
    * @return The integer that is the result of the multiplication
    */
-  public int op(int l, int r)
+  public double op(double l, double r)
     { return (l*r); }
 }
